delete mockitoTest, delete main, change salaryCalculator and salaryCalculatorTest
