+++ src/test/java/SalaryCalculatorTest.java
@@ -9,20 +9,27 @@ import java.io.IOException;
 
 public class SalaryCalculatorTest {
 
-    private Main main;
+    private SalaryCalculator salaryCalculator;
+
+    SalaryCalculator salaryCalculator0 = new SalaryCalculator("Andriy", "Vareg", 200, 172, 61, 100);
+    SalaryCalculator salaryCalculator1 = new SalaryCalculator("Vasyl", "Arial", 250, 172, 52, 125);
+    SalaryCalculator salaryCalculator2 = new SalaryCalculator("Oleh", "Serif", 50, 172, 9, 25);
+    SalaryCalculator salaryCalculator3 = new SalaryCalculator("Karl", "Aref", 150, 172, 12, 75);
+    SalaryCalculator salaryCalculator4 = new SalaryCalculator("Karolina", "Vagen", 100, 172, 0, 50);
+    SalaryCalculator salaryCalculator5 = new SalaryCalculator("Pavlo", "Agen", 300, 172, 150, 150);
 
 
     @Before
     public void setup() {
 
-        main = new Main();
+        salaryCalculator = new SalaryCalculator();
     }
 
 
     @Test
     public void salaryTest() {
 
-        int actual = main.salary(main.salaryCalculator0.getSalaryPerHour(), main.salaryCalculator0.getHoursWork(), main.salaryCalculator0.getHoursHospital(), main.salaryCalculator0.getHourHospitalSalary());
+        int actual = salaryCalculator.salary(salaryCalculator0.getSalaryPerHour(), salaryCalculator0.getHoursWork(), salaryCalculator0.getHoursHospital(), salaryCalculator0.getHourHospitalSalary());
         int expected = 20200;
 
         Assert.assertEquals("It is not true", expected, actual);
@@ -32,9 +39,9 @@ public class SalaryCalculatorTest {
     @Test
     public void hospitalSalaryTest() throws IOException {
 
-        main.salaryCalculator5.setHoursHospital(10);
+        salaryCalculator5.setHoursHospital(10);
 
-        int actual = main.hospitalSalary(main.salaryCalculator5.getHoursHospital(), main.salaryCalculator5.getHourHospitalSalary());
+        int actual = salaryCalculator.hospitalSalary(salaryCalculator5.getHoursHospital(), salaryCalculator5.getHourHospitalSalary());
         int expected = 1500;
         Assert.assertEquals(expected, actual);
         Assert.assertTrue(actual > 1000);
@@ -43,7 +50,7 @@ public class SalaryCalculatorTest {
     @Ignore("I don't have a time")
     public void isWorkSalaryTest20Thous() {
 
-        boolean actual = main.isWorkSalary(main.salary(main.salaryCalculator0.getSalaryPerHour(), main.salaryCalculator0.getHoursWork(), main.salaryCalculator0.getHoursHospital(), main.salaryCalculator0.getHourHospitalSalary()));
+        boolean actual = salaryCalculator.isWorkSalary(salaryCalculator.salary(salaryCalculator0.getSalaryPerHour(), salaryCalculator0.getHoursWork(), salaryCalculator0.getHoursHospital(), salaryCalculator0.getHourHospitalSalary()));
         boolean expected = true;
 
         Assert.assertTrue(actual);
